core-to-geb: sort the identifiers topologically
